integration work

+++ src/edu/esprit/pi/gui/AffichageAlerteController.java
@@ -12,12 +12,12 @@ import com.jfoenix.controls.JFXHamburger;
 import com.jfoenix.controls.JFXSlider;
 import com.jfoenix.controls.JFXTextField;
 import com.jfoenix.transitions.hamburger.HamburgerBackArrowBasicTransition;
+import edu.esprit.pi.iservices.ControlledScreen;
 import edu.esprit.pi.iservices.IAlerteService;
 import edu.esprit.pi.models.Alerte;
 import edu.esprit.pi.models.User;
 import edu.esprit.pi.services.AlerteService;
 import edu.esprit.pi.technique.DataSource;
-
 import java.io.IOException;
 import java.net.URL;
 import java.sql.Connection;
@@ -26,6 +26,7 @@ import java.sql.PreparedStatement;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.ResourceBundle;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -39,6 +40,8 @@ import javafx.fxml.Initializable;
 import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.Label;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
@@ -48,13 +51,19 @@ import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
+import javafx.util.Duration;
+import tray.notification.NotificationType;
+import tray.notification.TrayNotification;
 
 /**
  * FXML Controller class
  *
  * @author Nacef Fethi
  */
-public class AffichageAlerteController implements Initializable {
+public class AffichageAlerteController implements Initializable,ControlledScreen {
+      ScreensController screen;
+      User u1 = new User(User.getIdd());
+      
         @FXML
     private TableView<Alerte> TableAlerte;
      @FXML
@@ -108,7 +117,17 @@ public class AffichageAlerteController implements Initializable {
    
        private Connection connection;
     private PreparedStatement ps;
-     IAlerteService alerteService= new AlerteService();
+     AlerteService alerteService= new AlerteService();
+  //   public static Alerte alerteSelectionne;
+
+//    public static Alerte getAlerteSelectionne() {
+//        return alerteSelectionne;
+//    }
+//
+//    public static void setAlerteSelectionne(Alerte alerteSelectionne) {
+//        AffichageAlerteController.alerteSelectionne = alerteSelectionne;
+//    }
+     
 
     public AffichageAlerteController() {
                         connection = DataSource.getInstance().getConnection();
@@ -123,7 +142,8 @@ public class AffichageAlerteController implements Initializable {
         CLDate.setCellValueFactory(new PropertyValueFactory<>("date"));
         CLHeure.setCellValueFactory(new PropertyValueFactory<>("heure")); 
           lstAlerte = new ArrayList<>();
-           User user=new User(2);
+          System.out.println("ell"+u1.getId());
+           User user=new User(User.idd);
              
           lstAlerte = alerteService.getAll(user.getId());
            for (Alerte alerte : lstAlerte) {
@@ -147,7 +167,7 @@ public class AffichageAlerteController implements Initializable {
         dateP.setVisible(true);
         SliderHeure.setVisible(true);
         btnModifier.setVisible(true);
-        btnSupprimer.setVisible(true);
+   
         
             
             Alerte p=TableAlerte.getItems().get(TableAlerte.getSelectionModel().getSelectedIndex());
@@ -164,7 +184,18 @@ public class AffichageAlerteController implements Initializable {
     
     }
 
-
+//@FXML
+//      private void rech(){
+//          TableAlerte.getItems().clear();
+//    User user=new User(2);
+//alerteService.rechercher(txtChercher.getText(),user.getId());
+//           
+//             SetCellTable();
+//              
+//       
+//      
+//      
+//      }
       @FXML
     void rechercherAlerte(KeyEvent event) throws IOException {
          
@@ -172,7 +203,7 @@ public class AffichageAlerteController implements Initializable {
         
         lstAlerte = new ArrayList<>();
         AlerteService s=new AlerteService();
-        User user=new User(2);
+        User user=new User(User.getIdd());
         
        lstAlerte=s.rechercher(txtChercher.getText(),user.getId());
        // System.out.println(lstAlerte);
@@ -191,29 +222,62 @@ public class AffichageAlerteController implements Initializable {
                @FXML
                
     private void modifierAlerte(ActionEvent event) throws Exception{
-         User user=new User(2);
+         User user=new User(10);
         int id=Integer.parseInt(LId.getText());
          System.out.println(id);
  Alerte alerte =new Alerte (id,txtLieuDepart.getText(),txtLieuArrivee.getText(),java.sql.Date.valueOf(dateP.getValue()),(int) SliderHeure.getValue(),user);
  alerteService.update(alerte);
 //SetCellTable();
+ // screen.loadScreen(ScreensFramework.screen6ID, ScreensFramework.screen6File);
+             //       screen.setScreen(ScreensFramework.screen6ID);
  Parent affichageAlerte = FXMLLoader.load(getClass().getResource("AffichageAlerte.fxml"));
      Scene sceneAffichage = new Scene(affichageAlerte);
      Stage stage=(Stage) ((Node) event.getSource()).getScene().getWindow();
         stage.setScene(sceneAffichage);
       stage.show();
+// try {
+//           AlerteController controller = new AlerteController();
+//           //  controller.initVariable(groupeSelectionne);
+//        FXMLLoader afficher = new FXMLLoader();
+//        afficher.setLocation(getClass().getResource("alerte.fxml"));
+//         Parent  afficherGroupe = afficher.load();
+//
+//        Scene sceneAffichage = new Scene(afficherGroupe);
+//      
+//     Stage stage = (Stage) ((Node) event.getSource()).getScene().getWindow();
+//
+//        stage.setScene(sceneAffichage);
+//                    stage.show();
+//
+//                } catch (IOException ex) {
+//                                        Logger.getLogger(AffichageAlerteController.class.getName()).log(Level.SEVERE, null, ex);
+//
+//                    Logger.getLogger(AffichageAlerteController.class.getName()).log(Level.SEVERE, null, ex);
+//                }
+//      
 
 
     }
      @FXML
     private void supprimerAlerte(ActionEvent event) throws Exception{
         int id=Integer.parseInt(LId.getText());
-         alerteService.delete(id);
+        
+             Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+alert.setTitle("Confirmation Dialog");
+alert.setHeaderText("Look, a Confirmation Dialog");
+alert.setContentText("Are you ok with this?");
+
+Optional<ButtonType> result = alert.showAndWait();
+if (result.get() == ButtonType.OK){
+    alerteService.delete(id);
+} 
     Parent affichageAlerte = FXMLLoader.load(getClass().getResource("AffichageAlerte.fxml"));
      Scene sceneAffichage = new Scene(affichageAlerte);
      Stage stage=(Stage) ((Node) event.getSource()).getScene().getWindow();
         stage.setScene(sceneAffichage);
       stage.show();
+ // screen.loadScreen(ScreensFramework.screen6ID, ScreensFramework.screen6File);
+               //     screen.setScreen(ScreensFramework.screen6ID);
     }
        @FXML
     private void creerAlerte(ActionEvent event) throws Exception{
@@ -221,9 +285,20 @@ public class AffichageAlerteController implements Initializable {
      Scene sceneAffichage = new Scene(affichageAlerte);
      Stage stage=(Stage) ((Node) event.getSource()).getScene().getWindow();
         stage.setScene(sceneAffichage);
-      stage.show();}
+      stage.show();
+    //  screen.loadScreen(ScreensFramework.screen5ID, ScreensFramework.screen5File);
+    //                screen.setScreen(ScreensFramework.screen5ID);
+
+    }
     @Override
     public void initialize(URL url, ResourceBundle rb) {
+//        if(!alerteService.comparerPublicationAlerte(alerteService.getAll()).isEmpty())
+//{NotificationType type = NotificationType.INFORMATION;
+//            TrayNotification tray = new TrayNotification();
+//            tray.setTitle("Veuillez voir la liste des annonces");
+//            tray.setMessage(" on a trouvee une annonce adequate a votre besoin ");
+//            tray.setNotificationType(type);
+//            tray.showAndDismiss(Duration.seconds(5));}
      SetCellTable();
         setCellValueFromTableToText();
           labelLDepart.setVisible(false);
@@ -235,7 +310,7 @@ public class AffichageAlerteController implements Initializable {
          dateP.setVisible(false);
         SliderHeure.setVisible(false);
             btnModifier.setVisible(false);
-        btnSupprimer.setVisible(false);
+      
     //  chercherAlerte();
                 
         try {
@@ -261,4 +336,10 @@ public class AffichageAlerteController implements Initializable {
         });
        
     }  
+
+    @Override
+    public void setScreenParent(ScreensController screenPage) {
+        screen=screenPage;
+    }
 }
+
